refactor: a liitle refactor of context components
