+++ src/components/ChangeForm.tsx
@@ -1,4 +1,4 @@
-import React, { ReactEventHandler, useState } from "react";
+import React, { useState } from "react";
 import Button from "./Button";
 
 interface IChangeFormProps{
