+++ src/hooks/useFetch.tsx
@@ -2,7 +2,7 @@ import { useState } from "react";
 import axios, { AxiosError } from "axios";
 import { IWeather, errData } from "../interfaces";
 
-const useFetch = (callback:any) => {
+const useFetch = (callback:(params?:any)=>void) => {
   const [error, setError] = useState<null | AxiosError | Error | string>(null);
   const [isLoading, setIsLoading] = useState<boolean>(false);
   const fetchData:any = async (params:any) => {
